ANDROID: sched/fair: add idle state filter to prefer_idle case
The CPU selection process for a prefer_idle task either minimizes or
maximizes the CPU capacity for idle CPUs depending on the task being
boosted or not.

Given that we are iterating through all CPUs, additionally filter the
choice by preferring a CPU in a more shallow idle state. This will
provide both a faster wake-up for the task and higher energy efficiency,
by allowing CPUs in deeper idle states to remain idle.

Change-Id: I6a4734af9a3308277361a2851f06e83fc8f174e4
Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>